+++ src/visual.ts
@@ -106,6 +106,7 @@ import { ITooltipServiceWrapper, createTooltipServiceWrapper } from "powerbi-vis
 
 // powerbi.extensibility.utils.formattingModel
 import { FormattingSettingsService } from "powerbi-visuals-utils-formattingmodel";
+import ISelectionManager = powerbi.extensibility.ISelectionManager;
 
 const ColumnDisplayName: string = "Visual_Column";
 
@@ -198,6 +199,7 @@ export class StreamGraph implements IVisual {
     private dataPointsContainer: Selection<BaseType, StreamGraphSeries, any, any>;
 
     private localizationManager: ILocalizationManager;
+    private selectionManager: ISelectionManager;
 
     private static formattingSettingsService: FormattingSettingsService;
     private static formattingSettings: StreamGraphSettingsModel;
@@ -246,7 +248,6 @@ export class StreamGraph implements IVisual {
         colorPalette: IColorPalette,
         interactivityService: IInteractivityService<StreamGraphSeries>,
         visualHost: IVisualHost,
-        dataViews: DataView[]
     ): StreamData {
 
         if (!dataView
@@ -281,7 +282,7 @@ export class StreamGraph implements IVisual {
 
         const colorHelper: ColorHelper = new ColorHelper(colorPalette);
 
-        this.formattingSettings = this.formattingSettingsService.populateFormattingSettingsModel(StreamGraphSettingsModel, dataViews);
+        this.formattingSettings = this.formattingSettingsService.populateFormattingSettingsModel(StreamGraphSettingsModel, dataView);
         this.removeDisabledFormattingSettings(this.formattingSettings);
         const formattingSettings = this.formattingSettings;
         const fontSizeInPx: string = PixelConverter.fromPoint(formattingSettings.enableDataLabelsCardSettings.fontSize.value);
@@ -571,6 +572,7 @@ export class StreamGraph implements IVisual {
         this.visualHost = options.host;
         this.colorPalette = options.host.colorPalette;
         this.localizationManager = options.host.createLocalizationManager();
+        this.selectionManager = options.host.createSelectionManager();
         StreamGraph.formattingSettingsService = new FormattingSettingsService(this.localizationManager);
 
         const element: HTMLElement = options.element;
@@ -582,6 +584,8 @@ export class StreamGraph implements IVisual {
         this.svg = select(element)
             .append("svg")
             .classed(StreamGraph.VisualClassName, true);
+
+        this.handleContextMenu();
 
         this.clearCatcher = appendClearCatcher(this.svg);
 
@@ -637,7 +641,6 @@ export class StreamGraph implements IVisual {
             this.colorPalette,
             this.interactivityService,
             this.visualHost,
-            options.dataViews
         );
 
         if (!this.data
@@ -1259,7 +1262,9 @@ export class StreamGraph implements IVisual {
             dataLabelUtils.cleanDataLabels(this.svg);
         }
 
-        return selectionMerged        .attr("focusable", true);
+        selectionMerged.attr("focusable", true);
+
+        return selectionMerged;
     }
 
     private localizeLegendOrientationDropdown(enableLegendCardSettings : EnableLegendCardSettings)
@@ -1391,5 +1396,21 @@ export class StreamGraph implements IVisual {
 
     public getFormattingModel(): powerbi.visuals.FormattingModel {
         return StreamGraph.formattingSettingsService.buildFormattingModel(StreamGraph.formattingSettings);
+    }
+
+    private handleContextMenu() {
+        this.svg.on("contextmenu", (event) => {
+            const emptySelection = {
+                "measures": [],
+                "dataMap": {}
+            };
+
+            this.selectionManager.showContextMenu(emptySelection, {
+                x: event.clientX,
+                y: event.clientY
+            })
+
+            event.preventDefault();
+        })
     }
 }